+++ crt_portal/cts_forms/filters.py
@@ -2,10 +2,11 @@
 # provided they are valid filterable model properties.
 import re
 import urllib.parse
-from datetime import datetime
+from datetime import date, datetime, timedelta
 from django.core.validators import ValidationError
 
-from django.db.models import Count, Min
+from django.db.models import ExpressionWrapper, Count, Min, F, Func, Value, CharField, DateField, IntegerField
+from django.db.models.functions import ExtractDay, ExtractMonth, ExtractYear, Concat, Cast
 from django.contrib.postgres.search import SearchQuery
 from django.db import connection
 from django.http.request import QueryDict, MultiValueDict
@@ -207,6 +208,10 @@ def report_filter(querydict):
         elif field_options == 'disposition_status':
             disposition_status = querydict.getlist(field)[0]
             today = datetime.today().date()
+            qs = qs.annotate(retention_year = F('retention_schedule__retention_years'),
+                            expiration_year=F('retention_year') + ExtractYear('closed_date'),
+                            expiration_date=Cast(Concat(F('expiration_year'), Value('-'), ExtractMonth('closed_date'), Value('-'), ExtractDay('closed_date'), output_field=CharField()), output_field=DateField()),
+                            eligible_date=ExpressionWrapper(F('expiration_date') - timedelta(days=30), output_field=DateField()))
             if disposition_status == 'past':
                 kwargs[f'expiration_date__lt'] = today
             if disposition_status == 'eligible':
