+++ crt_portal/cts_forms/templates/forms/complaint_view/dashboard/index.html
@@ -28,6 +28,7 @@
 {% block page_js %}
 {{ block.super }}
 <script src="{% static 'js/dashboard_quick_view.min.js' %}"></script>
+<script src="{% static 'js/filters.min.js' %}"></script>
 <script src="{% static 'js/dashboard_view_filters.min.js' %}"></script>
 <script src="{% static 'js/ga_util.min.js' %}"></script>
 {% endblock %}
