+++ crt_portal/cts_forms/templates/forms/complaint_view/disposition/filter-controls.html
@@ -4,7 +4,7 @@
   action="/form/disposition"
   novalidate
 >
-    <button class="usa-button display-none" type="submit" id="apply-sort-page-button">
+    <button class="usa-button display-none" type="submit" id="apply-filters-button">
         Apply pagination
     </button>
 </form>
