+++ crt_portal/cts_forms/templates/forms/complaint_view/disposition/index.html
@@ -46,6 +46,7 @@
 <script src="{% static 'js/ga_util.min.js' %}"></script>
 <script src="{% static 'vendor/js.cookie.min.js' %}"></script>
 <script src="{% static 'js/complaint_quick_view.min.js' %}"></script>
+<script src="{% static 'js/filters.min.js' %}"></script>
 <script type="module" src="{% static 'js/disposition_pagination.min.js' %}"></script>
 <script src="{% static 'js/constant_writer.min.js'%}"></script>
 <script src="{% static 'js/complaint_actions.min.js' %}"></script>
