+++ crt_portal/cts_forms/templates/forms/complaint_view/index/grouped_index.html
@@ -79,6 +79,7 @@
 {{ block.super }}
 <script src="{% static 'vendor/js.cookie.min.js' %}"></script>
 <script src="{% static 'js/complaint_quick_view.min.js' %}"></script>
+<script type="module" src="{% static 'js/filters.min.js' %}"></script>
 <script type="module" src="{% static 'js/complaint_view_filters.min.js' %}"></script>
 <script src="{% static 'js/complaint_actions.min.js' %}"></script>
 <script src="{% static 'js/dashboard_view_all.min.js'%}"></script>
